mypt-ho-profile-api/requirements.txt
django
djangorestframework
markdown
django-filter
pyjwt
mysqlclient
pycryptodome
pycryptodomex
numpy
pandas
redis
python-decouple
requests
django-cors-headers
openpyxl
kafka-python